+++ src/components/TodoItem/index.jsx
@@ -1,8 +1,6 @@
-import { updateTodo } from 'apis/todoApi';
-import { useInput } from 'hooks/useInput';
-import { useEffect, useState } from 'react';
+import { useTodoItem } from './hook';
 
-export const TodoListItem = ({
+export const TodoItem = ({
   id,
   todo,
   isCompleted,
@@ -11,26 +9,13 @@ export const TodoListItem = ({
   setIsUpdated,
   isUpdated,
 }) => {
-  const { value: todoValue, onChange: onTodoChange } = useInput(todo);
-  const [isEdit, setIsEdit] = useState(false);
-  const [updatedTodoId, setUpdatedTodoId] = useState(null);
-
-  const changeEditState = () => {
-    setIsEdit((prev) => !prev);
-  };
-
-  const handleUpdateTodo = async () => {
-    await updateTodo(id, { todo: todoValue, isCompleted });
-    setIsUpdated(true);
-    setUpdatedTodoId(id);
-  };
-
-  useEffect(() => {
-    if (!isUpdated && updatedTodoId === id) {
-      setIsEdit(false);
-      setUpdatedTodoId(null);
-    }
-  }, [isUpdated]);
+  const { isEdit, onTodoChange, todoValue, handleUpdateTodo, changeEditState } = useTodoItem(
+    id,
+    setIsUpdated,
+    isUpdated,
+    todo,
+    isCompleted,
+  );
 
   return (
     <li>
